control-service: cleanup kubernetesService class
