+++ projects/control-service/projects/pipelines_control_service/src/test/java/com/vmware/taurus/datajobs/MockKubernetes.java
@@ -44,7 +44,7 @@ public class MockKubernetes {
   public DataJobsKubernetesService mockDataJobsKubernetesService()
       throws ApiException, IOException, InterruptedException {
     DataJobsKubernetesService mock = mock(DataJobsKubernetesService.class);
-    mockKubernetesService(mock);
+    mockDataJobsKubernetesService(mock);
     return mock;
   }
 
@@ -65,74 +65,119 @@ public class MockKubernetes {
     return new SyncTaskExecutor();
   }
 
+
+  private void mockKubernetesService(KubernetesService mock)
+          throws ApiException, IOException, InterruptedException {
+    // By defautl beans are singleton scoped so we are sure this will be called once
+    // hence it's safe to keep the variables here isntead of static.
+    final Map<String, Map<String, byte[]>> secrets = new ConcurrentHashMap<>();
+
+    final Map<String, InvocationOnMock> jobs = new ConcurrentHashMap<>();
+
+    when(mock.getSecretData(any()))
+            .thenAnswer(inv -> secrets.getOrDefault(inv.getArgument(0), Collections.emptyMap()));
+    doAnswer(answer(secrets::put)).when(mock).saveSecretData(any(), any());
+    doAnswer(inv -> secrets.remove(inv.getArgument(0))).when(mock).removeSecretData(any());
+
+    doAnswer(inv -> jobs.put(inv.getArgument(0), inv))
+            .when(mock)
+            .createJob(
+                    anyString(),
+                    anyString(),
+                    anyBoolean(),
+                    anyBoolean(),
+                    any(),
+                    any(),
+                    any(),
+                    any(),
+                    anyString(),
+                    any(),
+                    any(),
+                    anyLong(),
+                    anyLong(),
+                    anyLong(),
+                    anyString(),
+                    anyString());
+    doAnswer(inv -> jobs.remove(inv.getArgument(0))).when(mock).deleteJob(anyString());
+
+    doAnswer(
+            inv -> {
+              String jobName = inv.getArgument(0);
+              if (jobs.containsKey(jobName)) {
+                if (jobName.startsWith("failure-")) {
+                  return new KubernetesService.JobStatusCondition(
+                          false, "Status", "Job name starts with 'failure-'", "", 0);
+                } else {
+                  return new KubernetesService.JobStatusCondition(true, "Status", "", "", 0);
+                }
+              }
+              return new KubernetesService.JobStatusCondition(false, null, "No such job", "", 0);
+            })
+            .when(mock)
+            .watchJob(anyString(), anyInt(), any());
+
+    doAnswer(inv -> "logs").when(mock).getJobLogs(anyString(), anyInt());
+  }
   /**
    * Mocks interactions with KubernetesService as much as necessary for unit testing purpose.
    *
    * <p>NOTES: If job name starts with 'failure-' (e.g failure-my-job) - then Job status will be
    * fail otherwise it's success.
    */
-  private void mockKubernetesService(KubernetesService mock)
+  private void mockDataJobsKubernetesService(DataJobsKubernetesService mock)
       throws ApiException, IOException, InterruptedException {
-    // By defautl beans are singleton scoped so we are sure this will be called once
-    // hence it's safe to keep the variables here isntead of static.
-    final Map<String, Map<String, byte[]>> secrets = new ConcurrentHashMap<>();
+    mockKubernetesService(mock);
     final Map<String, InvocationOnMock> crons = new ConcurrentHashMap<>();
-    final Map<String, InvocationOnMock> jobs = new ConcurrentHashMap<>();
-
-    when(mock.getSecretData(any()))
-        .thenAnswer(inv -> secrets.getOrDefault(inv.getArgument(0), Collections.emptyMap()));
-    doAnswer(answer(secrets::put)).when(mock).saveSecretData(any(), any());
-    doAnswer(inv -> secrets.remove(inv.getArgument(0))).when(mock).removeSecretData(any());
 
     doAnswer(inv -> crons.put(inv.getArgument(0), inv))
-        .when(mock)
-        .createCronJob(
-            anyString(),
-            anyString(),
-            any(),
-            anyString(),
-            anyBoolean(),
-            any(),
-            any(),
-            any(),
-            any(),
-            any(),
-            any(),
-            any());
+            .when(mock)
+            .createCronJob(
+                    anyString(),
+                    anyString(),
+                    any(),
+                    anyString(),
+                    anyBoolean(),
+                    any(),
+                    any(),
+                    any(),
+                    any(),
+                    any(),
+                    any(),
+                    any());
     doAnswer(inv -> crons.put(inv.getArgument(0), inv))
-        .when(mock)
-        .createCronJob(
-            anyString(),
-            anyString(),
-            any(),
-            anyString(),
-            anyBoolean(),
-            any(),
-            any(),
-            any(),
-            any(),
-            any(),
-            any(),
-            any(),
-            any(),
-            any(),
-            any(),
-            anyList());
+            .when(mock)
+            .createCronJob(
+                    anyString(),
+                    anyString(),
+                    any(),
+                    anyString(),
+                    anyBoolean(),
+                    any(),
+                    any(),
+                    any(),
+                    any(),
+                    any(),
+                    any(),
+                    any(),
+                    any(),
+                    any(),
+                    any(),
+                    anyList());
     doAnswer(inv -> crons.put(inv.getArgument(0), inv))
-        .when(mock)
-        .updateCronJob(
-            anyString(),
-            anyString(),
-            any(),
-            anyString(),
-            anyBoolean(),
-            any(),
-            any(),
-            any(),
-            any(),
-            any(),
-            any(),
-            any());
+            .when(mock)
+            .updateCronJob(
+                    anyString(),
+                    anyString(),
+                    any(),
+                    anyString(),
+                    anyBoolean(),
+                    any(),
+                    any(),
+                    any(),
+                    any(),
+                    any(),
+                    any(),
+                    any());
     doAnswer(inv -> crons.keySet()).when(mock).listCronJobs();
     doAnswer(inv -> crons.remove(inv.getArgument(0))).when(mock).deleteCronJob(anyString());
     doAnswer(
@@ -147,47 +192,7 @@ public class MockKubernetes {
               }
               return Optional.ofNullable(deployment);
             })
-        .when(mock)
-        .readCronJob(anyString());
-
-    doAnswer(inv -> jobs.put(inv.getArgument(0), inv))
-        .when(mock)
-        .createJob(
-            anyString(),
-            anyString(),
-            anyBoolean(),
-            anyBoolean(),
-            any(),
-            any(),
-            any(),
-            any(),
-            anyString(),
-            any(),
-            any(),
-            anyLong(),
-            anyLong(),
-            anyLong(),
-            anyString(),
-            anyString());
-    doAnswer(inv -> jobs.keySet()).when(mock).listCronJobs();
-    doAnswer(inv -> jobs.remove(inv.getArgument(0))).when(mock).deleteJob(anyString());
-
-    doAnswer(
-            inv -> {
-              String jobName = inv.getArgument(0);
-              if (jobs.containsKey(jobName)) {
-                if (jobName.startsWith("failure-")) {
-                  return new KubernetesService.JobStatusCondition(
-                      false, "Status", "Job name starts with 'failure-'", "", 0);
-                } else {
-                  return new KubernetesService.JobStatusCondition(true, "Status", "", "", 0);
-                }
-              }
-              return new KubernetesService.JobStatusCondition(false, null, "No such job", "", 0);
-            })
-        .when(mock)
-        .watchJob(anyString(), anyInt(), any());
-
-    doAnswer(inv -> "logs").when(mock).getJobLogs(anyString(), anyInt());
+            .when(mock)
+            .readCronJob(anyString());
   }
 }
