+++ projects/control-service/projects/pipelines_control_service/src/test/java/com/vmware/taurus/service/KubernetesServiceStartNewCronJobExecutionTest.java
@@ -5,6 +5,7 @@
 
 package com.vmware.taurus.service;
 
+import com.vmware.taurus.service.kubernetes.DataJobsKubernetesService;
 import com.vmware.taurus.service.model.JobAnnotation;
 import com.vmware.taurus.service.model.JobEnvVar;
 import io.kubernetes.client.openapi.ApiException;
@@ -26,7 +27,7 @@ public class KubernetesServiceStartNewCronJobExecutionTest {
   public void testStartNewCronJobExecution_nullCronJobDefinition_shouldThrowException()
       throws ApiException {
     String jobName = "test-job";
-    KubernetesService kubernetesService = mockKubernetesService(jobName, null);
+    var kubernetesService = mockKubernetesService(jobName, null);
 
     Exception exception =
         Assertions.assertThrows(
@@ -54,7 +55,7 @@ public class KubernetesServiceStartNewCronJobExecutionTest {
             .spec(
                 new V1beta1CronJobSpec()
                     .jobTemplate(new V1beta1JobTemplateSpec().spec(new V1JobSpec())));
-    KubernetesService kubernetesService = mockKubernetesService(jobName, expectedResult);
+    var kubernetesService = mockKubernetesService(jobName, expectedResult);
 
     Exception exception =
         Assertions.assertThrows(
@@ -95,7 +96,7 @@ public class KubernetesServiceStartNewCronJobExecutionTest {
                                                     .addContainersItem(
                                                         new V1Container().addEnvItem(testEnv)))))));
 
-    KubernetesService kubernetesService = mockKubernetesService(jobName, expectedResult);
+    var kubernetesService = mockKubernetesService(jobName, expectedResult);
     kubernetesService.startNewCronJobExecution(
         jobName,
         executionId,
@@ -172,7 +173,7 @@ public class KubernetesServiceStartNewCronJobExecutionTest {
                                                 new V1PodSpec()
                                                     .addContainersItem(new V1Container()))))));
 
-    KubernetesService kubernetesService = mockKubernetesService(jobName, expectedResult);
+    var kubernetesService = mockKubernetesService(jobName, expectedResult);
     kubernetesService.startNewCronJobExecution(
         jobName,
         executionId,
@@ -202,9 +203,9 @@ public class KubernetesServiceStartNewCronJobExecutionTest {
     Assertions.assertEquals(expectedAnnotations, annotationsArgumentCaptor.getValue());
   }
 
-  private KubernetesService mockKubernetesService(String jobName, V1beta1CronJob result)
+  private DataJobsKubernetesService mockKubernetesService(String jobName, V1beta1CronJob result)
       throws ApiException {
-    KubernetesService kubernetesService = Mockito.mock(KubernetesService.class);
+    var kubernetesService = Mockito.mock(DataJobsKubernetesService.class);
     Mockito.doCallRealMethod()
         .when(kubernetesService)
         .startNewCronJobExecution(
